requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
alsa-lib=1.2.12=h4ab18f5_0
brotli=1.1.0=hb9d3cd8_2
brotli-bin=1.1.0=hb9d3cd8_2
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.8.30=hbcca054_0
cairo=1.18.0=hebfffa5_3
certifi=2024.8.30=pyhd8ed1ab_0
contfrac=1.0.0=pypi_0
contourpy=1.3.0=py310h3788b33_1
cuda-quantum=0.8.0=pypi_0
cycler=0.12.1=pyhd8ed1ab_0
dbus=1.13.6=h5008d03_3
double-conversion=3.3.0=h59595ed_0
expat=2.6.3=h5888daf_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_2
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.53.1=py310ha75aee5_1
freetype=2.12.1=h267a509_2
graphite2=1.3.13=h59595ed_1003
harfbuzz=9.0.0=hda332d3_1
icu=75.1=he02047a_0
idna=3.8=pypi_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.7=py310h3788b33_0
krb5=1.21.3=h659f571_0
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.38=h1181459_1
lerc=4.0.0=h27087fc_0
libblas=3.9.0=23_linux64_openblas
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcblas=3.9.0=23_linux64_openblas
libclang-cpp18.1=18.1.8=default_hf981a13_4
libclang13=18.1.8=default_h9def88c_4
libcups=2.3.3=h4637d8d_4
libdeflate=1.21=h4bc722e_0
libdrm=2.4.123=hb9d3cd8_0
libedit=3.1.20191231=he28a2e2_2
libegl=1.7.0=ha4b6fd6_0
libexpat=2.6.3=h5888daf_0
libffi=3.4.4=h6a678d5_1
libgcc=14.1.0=h77fa898_1
libgcc-ng=14.1.0=h69a702a_1
libgfortran=14.1.0=h69a702a_1
libgfortran-ng=14.1.0=h69a702a_1
libgfortran5=14.1.0=hc5f4f2c_1
libgl=1.7.0=ha4b6fd6_0
libglib=2.80.3=h315aac3_2
libglvnd=1.7.0=ha4b6fd6_0
libglx=1.7.0=ha4b6fd6_0
libgomp=14.1.0=h77fa898_1
libiconv=1.17=hd590300_2
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=23_linux64_openblas
libllvm18=18.1.8=h8b73ec9_2
libnsl=2.0.1=hd590300_0
libopenblas=0.3.27=pthreads_hac2b453_1
libpciaccess=0.18=hd590300_0
libpng=1.6.44=hadc24fc_0
libpq=16.4=h2d7952a_1
libsqlite=3.46.1=hadc24fc_0
libstdcxx=14.1.0=hc0a3c3a_1
libstdcxx-ng=14.1.0=h4852527_1
libtiff=4.6.0=h46a8edc_4
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.4.0=hd590300_0
libxcb=1.16=hb9d3cd8_1
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.7.0=h2c5496b_1
libxml2=2.12.7=he7c6b58_4
libxslt=1.1.39=h76b75d6_0
libzlib=1.3.1=h4ab18f5_1
matplotlib=3.9.2=py310hff52083_1
matplotlib-base=3.9.2=py310h68603db_1
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=9.0.1=h70512c7_0
mysql-libs=9.0.1=ha479ceb_0
ncurses=6.4=h6a678d5_0
numpy=2.1.1=py310hd6e36ab_0
openjpeg=2.5.2=h488ebb8_0
openssl=3.3.2=hb9d3cd8_0
packaging=24.1=pyhd8ed1ab_0
pcre2=10.44=hba22ea6_2
pillow=10.4.0=py310he228d35_1
pip=24.2=py310h06a4308_0
pixman=0.43.2=h59595ed_0
pthread-stubs=0.4=h36c2ea0_1001
pyparsing=3.1.4=pyhd8ed1ab_0
pyside6=6.7.2=py310hfd10a26_3
python=3.10.13=hd12c33a_1_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.10=5_cp310
qhull=2020.2=h434a139_5
qt6-main=6.7.2=hb12f9c5_5
readline=8.2=h5eee18b_0
setuptools=72.1.0=py310h06a4308_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.32.3=hcee41ef_1
tk=8.6.13=noxft_h4845f30_101
tornado=6.4.1=py310ha75aee5_1
tzdata=2024a=h04d1e81_0
unicodedata2=15.1.0=py310h2372a71_0
urllib3=2.2.3=pypi_0
wayland=1.23.1=h3e06ad9_0
wheel=0.44.0=py310h06a4308_0
xcb-util=0.4.1=hb711507_2
xcb-util-cursor=0.1.5=hb9d3cd8_0
xcb-util-image=0.4.0=hb711507_2
xcb-util-keysyms=0.4.1=hb711507_0
xcb-util-renderutil=0.3.10=hb711507_0
xcb-util-wm=0.4.2=hb711507_0
xkeyboard-config=2.42=h4ab18f5_0
xorg-fixesproto=5.0=h7f98852_1002
xorg-inputproto=2.3.2=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.1.1=hd590300_0
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.9=hb711507_1
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxi=1.7.10=h4bc722e_1
xorg-libxrender=0.9.11=hd590300_0
xorg-libxtst=1.2.5=h4bc722e_0
xorg-libxxf86vm=1.1.5=hb9d3cd8_2
xorg-recordproto=1.14.2=h7f98852_1002
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xz=5.4.6=h5eee18b_1
zlib=1.3.1=h4ab18f5_1
zstd=1.5.6=ha6fb4c9_0
